refactor(storage): wrap postgres configuration in an enclosing factory object
