+++ storage/build.gradle.kts
@@ -2,10 +2,15 @@ dependencies {
     api(project(":application"))
     with(libs) {
         implementation(postgresql)
-        implementation(exposed.core)
-        implementation(exposed.jdbc)
+        api(exposed.core)
+        api(exposed.jdbc)
         implementation(hikaricp)
     }
+}
+
+dockerCompose {
+    val postgresService = "postgres-db"
+    startedServices = listOf(postgresService)
 }
 
 dockerCompose.isRequiredBy(tasks.test)
