+++ storage/src/test/kotlin/io/github/positionpal/notification/storage/tokens/PostgresGroupsRepositoryTest.kt
@@ -2,7 +2,7 @@ package io.github.positionpal.notification.storage.tokens
 
 import io.github.positionpal.entities.GroupId
 import io.github.positionpal.entities.UserId
-import io.github.positionpal.notification.storage.PostgresConnectionFactory
+import io.github.positionpal.notification.storage.Postgres
 import io.github.positionpal.notification.storage.groups.PostgresGroupsRepository
 import io.kotest.common.runBlocking
 import io.kotest.core.spec.style.WordSpec
@@ -11,7 +11,7 @@ import io.kotest.matchers.shouldBe
 class PostgresGroupsRepositoryTest : WordSpec({
 
     beforeSpec {
-        connectionFactory.connect()
+        Postgres(localPostgresConfiguration).connect().onFailure { error("Failure in DB connection!") }
     }
 
     "PostgresUsersTokensRepository" When {
@@ -72,13 +72,6 @@ class PostgresGroupsRepositoryTest : WordSpec({
     }
 }) {
     private companion object {
-        private val connectionFactory = PostgresConnectionFactory(
-            databaseName = "notifications_service",
-            username = "postgres",
-            password = "postgres",
-            host = "localhost",
-            port = 5432,
-        )
         private val repository = PostgresGroupsRepository()
         private val groups = setOf(
             GroupId.create("astro") to setOf(UserId.create("Luca"), UserId.create("Greg")),
